refactor(auth): convert useWorkspaceAuth to Pinia store and fix auth issues
- Convert useWorkspaceAuth composable to workspaceAuthStore for proper
  global singleton state management
- Replace VueUse useTimeoutFn with native setTimeout for timer management
- Remove getCurrentInstance() guard (stores don't need component context)
- Extract WORKSPACE_STORAGE_KEYS to workspaceConstants.ts to avoid
  circular dependencies
- Keep useWorkspaceAuth() as backward-compatible wrapper

Fix session cookie error handling:
- createSession: log warnings instead of throwing (graceful degradation)
- deleteSession: log warnings instead of throwing (logout always succeeds)

Fix onUserLogout firing on initial mount:
- Check prevUser && !user to only fire on logout transition

Amp-Thread-ID: https://ampcode.com/threads/T-019bc3ba-76df-764f-8e1e-73864ff0b477
