+++ jwt-firebase-schema.md
@@ -1,0 +1,9 @@
+The key is like this:
+
+```
+firebase:authUser:AIzaSyC2-fomLqgCjb7ELwta1I9cEarPK8ziTGs:[DEFAULT]
+```
+
+The value is like this:
+
+{"uid":"dVUW1cny9fXRKpgfLB3vf6BZwZ12","email":"cbyrne@drip.art","emailVerified":true,"displayName":"Christian Byrne","isAnonymous":false,"photoURL":"https://lh3.googleusercontent.com/a/ACg8ocK0tdTpWBCTfdb9OEsJb5RSTzU2mOAsotVHUjByWSgTTPjqUA=s96-c","providerData":[{"providerId":"google.com","uid":"115296307001947560103","displayName":"Christian Byrne","email":"cbyrne@comfy.org","phoneNumber":null,"photoURL":"https://lh3.googleusercontent.com/a/ACg8ocK0tdTpWBCTfdb9OEsJb5RSTzU2mOAsotVHUjByWSgTTPjqUA=s96-c"}],"stsTokenManager":{"refreshToken":"AMf-vBxyNvwXiAZmBzib6fQW1Cl9c-b0VWckU3G_LZs9bJ17biKt5qqqzV6yLtR5ph7WZmR-oaxM50n5FzzWujgSS3x9qXqHW-rKAK9eOaRjV3WylP5JZd2xcsBiJ_YCsN3UksUGAMuE0FSSLJX9HY04x76VfWt8sVQzAN3h06nBYTr3T7D5D-vFRyUGSeBVvoADlE4bTjjRegd7OVHFr4H9lO2bCc2B-O8prZ9MMwlS8kvWcPLluQrd8ZIYFf8aj7MoDv8r8CtzeezhGFIx9vgCLXbx4M2KYBzP2rlS-4dz46vbvVIgErguvicu_Mv-yuDBlrZKHkri5ohLcqOnpX0-ulDvy3_3ZNGc23DrhzpM_6jUXl7mXgKbhiYRuNmHGZtWyQwygxz3Ai8y6cOBHvT0F_xPHly9nzGskejN9auo0VZvKIbaJF4","accessToken":"eyJhbGciOiJSUzI1NiIsImtpZCI6ImQ4Mjg5MmZhMzJlY2QxM2E0ZTBhZWZlNjI4ZGQ5YWFlM2FiYThlMWUiLCJ0eXAiOiJKV1QifQ.eyJuYW1lIjoiQ2hyaXN0aWFuIEJ5cm5lIiwicGljdHVyZSI6Imh0dHBzOi8vbGgzLmdvb2dsZXVzZXJjb250ZW50LmNvbS9hL0FDZzhvY0swdGRUcFdCQ1RmZGI5T0VzSmI1UlNUelUybU9Bc290VkhVakJ5V1NnVFRQanFVQT1zOTYtYyIsImlzcyI6Imh0dHBzOi8vc2VjdXJldG9rZW4uZ29vZ2xlLmNvbS9kcmVhbWJvb3RoeSIsImF1ZCI6ImRyZWFtYm9vdGh5IiwiYXV0aF90aW1lIjoxNzY4MzQyNTI4LCJ1c2VyX2lkIjoiZFZVVzFjbnk5ZlhSS3BnZkxCM3ZmNkJad1oxMiIsInN1YiI6ImRWVVcxY255OWZYUktwZ2ZMQjN2ZjZCWndaMTIiLCJpYXQiOjE3NjgzNDI1MjgsImV4cCI6MTc2ODM0NjEyOCwiZW1haWwiOiJjYnlybmVAZHJpcC5hcnQiLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiZmlyZWJhc2UiOnsiaWRlbnRpdGllcyI6eyJnb29nbGUuY29tIjpbIjExNTI5NjMwNzAwMTk0NzU2MDEwMyJdLCJlbWFpbCI6WyJjYnlybmVAZHJpcC5hcnQiXX0sInNpZ25faW5fcHJvdmlkZXIiOiJnb29nbGUuY29tIn19.p5tO1ZUXp_528gL2NWP5m26JYsCcth5fFpfcDnkOnQU7hlMVebcWq5I-riciFmzbqeE77lNaIEGM272W6YoTG_ULu0eo0i31DPoEr1B0xOBNxRDzU7KHKTrkddOzetiQrYrx4Gqfza5xiSDAPJzwUK_ihmD1OWTKXgducbDtVwxYIodfPMGyoprrlWFzF0qJAx06VfdzfATUl-JxV3t8ZW9aFrzdnYcDerRn9Mj3FJqREGVaMVMQGBQWRw0-lQKsk-uatjwKHFdNy6sQTV8XWeQ51HOJWQ9tT6H0bS7EFwKLiBYiQbrxyJp9ZeK_ajIBHFafB2PE29RnKKHnFl7Giw","expirationTime":1768346128532},"createdAt":"1744833024270","lastLoginAt":"1768342528496","apiKey":"AIzaSyC2-fomLqgCjb7ELwta1I9cEarPK8ziTGs","appName":"[DEFAULT]"}
