+++ src/platform/auth/workspace/useWorkspaceAuth.ts
@@ -1,332 +1,19 @@
-import { useTimeoutFn } from '@vueuse/core'
-import { computed, getCurrentInstance, ref, shallowRef } from 'vue'
-import { z } from 'zod'
-import { fromZodError } from 'zod-validation-error'
+import { storeToRefs } from 'pinia'
 
-import { t } from '@/i18n'
-import { remoteConfig } from '@/platform/remoteConfig/remoteConfig'
-import { api } from '@/scripts/api'
-import { useFirebaseAuthStore } from '@/stores/firebaseAuthStore'
-import type { AuthHeader } from '@/types/authTypes'
-import type { WorkspaceWithRole } from './workspaceTypes'
+import { useWorkspaceAuthStore } from '@/stores/workspaceAuthStore'
 
-const WorkspaceWithRoleSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  type: z.enum(['personal', 'team']),
-  role: z.enum(['owner', 'member'])
-})
+export { WORKSPACE_STORAGE_KEYS } from './workspaceConstants'
+export { WorkspaceAuthError } from '@/stores/workspaceAuthStore'
 
-const WorkspaceTokenResponseSchema = z.object({
-  token: z.string(),
-  expires_at: z.string(),
-  workspace: z.object({
-    id: z.string(),
-    name: z.string(),
-    type: z.enum(['personal', 'team'])
-  }),
-  role: z.enum(['owner', 'member']),
-  permissions: z.array(z.string())
-})
-
-export const WORKSPACE_STORAGE_KEYS = {
-  CURRENT_WORKSPACE: 'Comfy.Workspace.Current',
-  TOKEN: 'Comfy.Workspace.Token',
-  EXPIRES_AT: 'Comfy.Workspace.ExpiresAt'
-} as const
-
-const TOKEN_REFRESH_BUFFER_MS = 5 * 60 * 1000
-
-export class WorkspaceAuthError extends Error {
-  constructor(
-    message: string,
-    public readonly code?: string
-  ) {
-    super(message)
-    this.name = 'WorkspaceAuthError'
-  }
-}
-
-/**
- * Composable for managing per-tab workspace authentication.
- *
- * **Must be called from a Vue component setup context** because it uses
- * VueUse's useTimeoutFn which requires lifecycle hooks for auto-cleanup.
- * Using this composable outside of a component setup function will cause errors.
- *
- * @throws {Error} If called outside of a Vue component setup context
- */
 export function useWorkspaceAuth() {
-  if (!getCurrentInstance()) {
-    throw new Error(
-      'useWorkspaceAuth must be called from within a Vue component setup function'
-    )
-  }
-
-  const firebaseAuthStore = useFirebaseAuthStore()
-
-  const currentWorkspace = shallowRef<WorkspaceWithRole | null>(null)
-  const workspaceToken = ref<string | null>(null)
-  const isLoading = ref(false)
-  const error = ref<Error | null>(null)
-
-  const refreshDelay = ref(0)
-  const { start: startRefreshTimer, stop: stopRefreshTimer } = useTimeoutFn(
-    () => {
-      void refreshToken()
-    },
-    refreshDelay,
-    { immediate: false }
-  )
-
-  const isAuthenticated = computed(
-    () => currentWorkspace.value !== null && workspaceToken.value !== null
-  )
-
-  function scheduleTokenRefresh(expiresAt: number): void {
-    stopRefreshTimer()
-    const now = Date.now()
-    const refreshAt = expiresAt - TOKEN_REFRESH_BUFFER_MS
-    refreshDelay.value = Math.max(0, refreshAt - now)
-
-    startRefreshTimer()
-  }
-
-  function persistToSession(
-    workspace: WorkspaceWithRole,
-    token: string,
-    expiresAt: number
-  ): void {
-    try {
-      sessionStorage.setItem(
-        WORKSPACE_STORAGE_KEYS.CURRENT_WORKSPACE,
-        JSON.stringify(workspace)
-      )
-      sessionStorage.setItem(WORKSPACE_STORAGE_KEYS.TOKEN, token)
-      sessionStorage.setItem(
-        WORKSPACE_STORAGE_KEYS.EXPIRES_AT,
-        expiresAt.toString()
-      )
-    } catch {
-      console.warn('Failed to persist workspace context to sessionStorage')
-    }
-  }
-
-  function clearSessionStorage(): void {
-    try {
-      sessionStorage.removeItem(WORKSPACE_STORAGE_KEYS.CURRENT_WORKSPACE)
-      sessionStorage.removeItem(WORKSPACE_STORAGE_KEYS.TOKEN)
-      sessionStorage.removeItem(WORKSPACE_STORAGE_KEYS.EXPIRES_AT)
-    } catch {
-      console.warn('Failed to clear workspace context from sessionStorage')
-    }
-  }
-
-  function initializeFromSession(): boolean {
-    if (!remoteConfig.value.team_workspaces_enabled) {
-      return false
-    }
-
-    try {
-      const workspaceJson = sessionStorage.getItem(
-        WORKSPACE_STORAGE_KEYS.CURRENT_WORKSPACE
-      )
-      const token = sessionStorage.getItem(WORKSPACE_STORAGE_KEYS.TOKEN)
-      const expiresAtStr = sessionStorage.getItem(
-        WORKSPACE_STORAGE_KEYS.EXPIRES_AT
-      )
-
-      if (!workspaceJson || !token || !expiresAtStr) {
-        return false
-      }
-
-      const expiresAt = parseInt(expiresAtStr, 10)
-      if (isNaN(expiresAt) || expiresAt <= Date.now()) {
-        clearSessionStorage()
-        return false
-      }
-
-      const parsedWorkspace = JSON.parse(workspaceJson)
-      const parseResult = WorkspaceWithRoleSchema.safeParse(parsedWorkspace)
-
-      if (!parseResult.success) {
-        clearSessionStorage()
-        return false
-      }
-
-      currentWorkspace.value = parseResult.data
-      workspaceToken.value = token
-      error.value = null
-
-      scheduleTokenRefresh(expiresAt)
-      return true
-    } catch {
-      clearSessionStorage()
-      return false
-    }
-  }
-
-  async function switchWorkspace(workspaceId: string): Promise<void> {
-    if (!remoteConfig.value.team_workspaces_enabled) {
-      return
-    }
-
-    isLoading.value = true
-    error.value = null
-
-    try {
-      const firebaseToken = await firebaseAuthStore.getIdToken()
-      if (!firebaseToken) {
-        throw new WorkspaceAuthError(
-          t('workspaceAuth.errors.notAuthenticated'),
-          'NOT_AUTHENTICATED'
-        )
-      }
-
-      const response = await fetch(api.apiURL('/auth/token'), {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${firebaseToken}`,
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ workspace_id: workspaceId })
-      })
-
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => ({}))
-        const message = errorData.message || response.statusText
-
-        if (response.status === 401) {
-          throw new WorkspaceAuthError(
-            t('workspaceAuth.errors.invalidFirebaseToken'),
-            'INVALID_FIREBASE_TOKEN'
-          )
-        }
-        if (response.status === 403) {
-          throw new WorkspaceAuthError(
-            t('workspaceAuth.errors.accessDenied'),
-            'ACCESS_DENIED'
-          )
-        }
-        if (response.status === 404) {
-          throw new WorkspaceAuthError(
-            t('workspaceAuth.errors.workspaceNotFound'),
-            'WORKSPACE_NOT_FOUND'
-          )
-        }
-
-        throw new WorkspaceAuthError(
-          t('workspaceAuth.errors.tokenExchangeFailed', { error: message }),
-          'TOKEN_EXCHANGE_FAILED'
-        )
-      }
-
-      const rawData = await response.json()
-      const parseResult = WorkspaceTokenResponseSchema.safeParse(rawData)
-
-      if (!parseResult.success) {
-        throw new WorkspaceAuthError(
-          t('workspaceAuth.errors.tokenExchangeFailed', {
-            error: fromZodError(parseResult.error).message
-          }),
-          'TOKEN_EXCHANGE_FAILED'
-        )
-      }
-
-      const data = parseResult.data
-      const expiresAt = new Date(data.expires_at).getTime()
-
-      if (isNaN(expiresAt)) {
-        throw new WorkspaceAuthError(
-          t('workspaceAuth.errors.tokenExchangeFailed', {
-            error: 'Invalid expiry timestamp'
-          }),
-          'TOKEN_EXCHANGE_FAILED'
-        )
-      }
-
-      const workspaceWithRole: WorkspaceWithRole = {
-        ...data.workspace,
-        role: data.role
-      }
-
-      currentWorkspace.value = workspaceWithRole
-      workspaceToken.value = data.token
-
-      persistToSession(workspaceWithRole, data.token, expiresAt)
-      scheduleTokenRefresh(expiresAt)
-    } catch (err) {
-      error.value = err instanceof Error ? err : new Error(String(err))
-      throw error.value
-    } finally {
-      isLoading.value = false
-    }
-  }
-
-  async function refreshToken(): Promise<void> {
-    if (!currentWorkspace.value) {
-      return
-    }
-
-    const workspaceId = currentWorkspace.value.id
-    const maxRetries = 3
-    const baseDelayMs = 1000
-
-    for (let attempt = 0; attempt <= maxRetries; attempt++) {
-      try {
-        await switchWorkspace(workspaceId)
-        return
-      } catch (err) {
-        const isAuthError = err instanceof WorkspaceAuthError
-
-        const isPermanentError =
-          isAuthError &&
-          (err.code === 'ACCESS_DENIED' ||
-            err.code === 'WORKSPACE_NOT_FOUND' ||
-            err.code === 'INVALID_FIREBASE_TOKEN' ||
-            err.code === 'NOT_AUTHENTICATED')
-
-        if (isPermanentError) {
-          console.error('Workspace access revoked or auth invalid:', err)
-          clearWorkspaceContext()
-          return
-        }
-
-        const isTransientError =
-          isAuthError && err.code === 'TOKEN_EXCHANGE_FAILED'
-
-        if (isTransientError && attempt < maxRetries) {
-          const delay = baseDelayMs * Math.pow(2, attempt)
-          console.warn(
-            `Token refresh failed (attempt ${attempt + 1}/${maxRetries + 1}), retrying in ${delay}ms:`,
-            err
-          )
-          await new Promise((resolve) => setTimeout(resolve, delay))
-          continue
-        }
-
-        console.error('Failed to refresh workspace token after retries:', err)
-        clearWorkspaceContext()
-      }
-    }
-  }
-
-  function getWorkspaceAuthHeader(): AuthHeader | null {
-    if (!workspaceToken.value) {
-      return null
-    }
-    return {
-      Authorization: `Bearer ${workspaceToken.value}`
-    }
-  }
-
-  function clearWorkspaceContext(): void {
-    stopRefreshTimer()
-    currentWorkspace.value = null
-    workspaceToken.value = null
-    error.value = null
-    clearSessionStorage()
-  }
+  const store = useWorkspaceAuthStore()
+  const {
+    currentWorkspace,
+    workspaceToken,
+    isLoading,
+    error,
+    isAuthenticated
+  } = storeToRefs(store)
 
   return {
     currentWorkspace,
@@ -334,10 +21,10 @@ export function useWorkspaceAuth() {
     isLoading,
     error,
     isAuthenticated,
-    initializeFromSession,
-    switchWorkspace,
-    refreshToken,
-    getWorkspaceAuthHeader,
-    clearWorkspaceContext
+    initializeFromSession: store.initializeFromSession,
+    switchWorkspace: store.switchWorkspace,
+    refreshToken: store.refreshToken,
+    getWorkspaceAuthHeader: store.getWorkspaceAuthHeader,
+    clearWorkspaceContext: store.clearWorkspaceContext
   }
 }
